# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Game backend API với PlayFab integration, anti-cheat validation, và talent system",
        "title": "Hollow Wilds Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gamefeel.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.1.0"
    },
    "host": "gamefeel-backend.fly.dev",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "security": [
                    {
                        "PlayFabToken": []
                    }
                ],
                "description": "Authenticate user with PlayFab session token and get JWT for API access",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login with PlayFab",
                "parameters": [
                    {
                        "type": "string",
                        "description": "PlayFab session token",
                        "name": "X-PlayFab-SessionToken",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Login request with PlayFab ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AuthRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful login with JWT token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Invalid PlayFab token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.APIError"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.AuthRequest": {
            "type": "object",
            "required": [
                "playfabId"
            ],
            "properties": {
                "displayName": {
                    "type": "string"
                },
                "playfabId": {
                    "type": "string"
                }
            }
        },
        "models.AuthResponse": {
            "type": "object",
            "properties": {
                "expiresIn": {
                    "description": "seconds",
                    "type": "integer"
                },
                "jwt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "banReason": {
                    "type": "string"
                },
                "bannedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "description": "Metadata",
                    "type": "string"
                },
                "deletedAt": {
                    "description": "GDPR",
                    "type": "string"
                },
                "diamonds": {
                    "type": "integer"
                },
                "displayName": {
                    "type": "string"
                },
                "facebookId": {
                    "description": "Social",
                    "type": "string"
                },
                "gold": {
                    "description": "Currency",
                    "type": "integer"
                },
                "googleId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isBanned": {
                    "description": "Flags",
                    "type": "boolean"
                },
                "lastLoginAt": {
                    "type": "string"
                },
                "lastPlayedAt": {
                    "type": "string"
                },
                "maxMapUnlocked": {
                    "description": "Progression",
                    "type": "integer"
                },
                "playfabId": {
                    "type": "string"
                },
                "totalPlayTimeSeconds": {
                    "type": "integer"
                },
                "totalStarsCollected": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "PlayFabToken": {
            "description": "PlayFab session token for authentication",
            "type": "apiKey",
            "name": "X-PlayFab-SessionToken",
            "in": "header"
        }
    }
}